backend/requirements.txt
fastapi
uvicorn
requests
pyyaml
sqlalchemy
pydantic
python-multipart
python-jose
passlib==1.7.4
bcrypt==3.2.2
websockets
apscheduler
psycopg2-binary
ldap3
pywinrm
paramiko
celery
redis
aiohttp
kombu
pytest
pytest-cov
